tidy(ui): organise canvas entity classes
