+++ invokeai/frontend/web/src/features/controlLayers/konva/CanvasStateApiModule.ts
@@ -2,8 +2,8 @@ import { $alt, $ctrl, $meta, $shift } from '@invoke-ai/ui-library';
 import type { Selector } from '@reduxjs/toolkit';
 import { addAppListener } from 'app/store/middleware/listenerMiddleware';
 import type { AppStore, RootState } from 'app/store/store';
-import type { CanvasEntityAdapterControlLayer } from 'features/controlLayers/konva/CanvasEntityAdapter/CanvasEntityAdapterControlLayer';
-import type { CanvasEntityAdapterRasterLayer } from 'features/controlLayers/konva/CanvasEntityAdapter/CanvasEntityAdapterRasterLayer';
+import type { CanvasEntityAdapterControlLayer } from 'features/controlLayers/konva/CanvasEntity/CanvasEntityAdapterControlLayer';
+import type { CanvasEntityAdapterRasterLayer } from 'features/controlLayers/konva/CanvasEntity/CanvasEntityAdapterRasterLayer';
 import type { CanvasManager } from 'features/controlLayers/konva/CanvasManager';
 import { CanvasModuleBase } from 'features/controlLayers/konva/CanvasModuleBase';
 import type { SubscriptionHandler } from 'features/controlLayers/konva/util';
@@ -43,7 +43,7 @@ import { queueApi } from 'services/api/endpoints/queue';
 import type { BatchConfig } from 'services/api/types';
 import { assert } from 'tsafe';
 
-import type { CanvasEntityAdapter } from './CanvasEntityAdapter/types';
+import type { CanvasEntityAdapter } from './CanvasEntity/types';
 
 export class CanvasStateApiModule extends CanvasModuleBase {
   readonly type = 'state_api';
